Bugfix timeformat for when is reset the stats for queue:
- Move the tests for convert_time_when_param inside a own file
- Fix the error for strftime time

Close #168
